Added ft_atoi documentation

+++ ft_ltoa_base.c
@@ -6,13 +6,13 @@
 /*   By: bvelasco <bvelasco@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 16:47:41 by bvelasco          #+#    #+#             */
-/*   Updated: 2023/01/15 02:24:40 by bvelasco         ###   ########.fr       */
+/*   Updated: 2023/01/16 19:39:46 by bvelasco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_ltoa_base(long lnbr, char *base)
+char	*ft_ltoa_base(long long lnbr, char *base)
 {
 	int		base_len;
 	char	*ret;
